patch move customziation class back to shared/codewhisperer
